requirements.txt
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
gspread==5.10.0
openai==1.35.0
pillow==10.0.0
pillow-heif==0.13.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0